+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Unreleased
 
+### Added
+
+### Changed
+
+- `CustomResource` should have plain `apiVersion` and `kind` properties (https://github.com/pulumi/pulumi-kubernetes/pull/3079)
+
+### Fixed
+
 ## 4.14.0 (June 28, 2024)
 
 ### Added